refactor(archestra-bench): extract shared core contract crate
archestra-bench-core (serde-only) owns the run.json/trajectory.jsonl/outcome/
layout contract that the harness writes and the analyzer reads:
- slug + cell_dir (the <env>/<task>__<lane> layout that drifted and broke before)
- Outcome, RunMeta, RolloutId, trajectory Event

Runner and analyzer both depend on it (deleting their private copies), so the
read/write contract can no longer silently diverge. Adds a runner-side round-trip
test pinning run.json/trajectory output to the core types.
