+++ archestra-bench/analyzer/Cargo.toml
@@ -2,12 +2,14 @@
 name = "trajectory-analyzer"
 version = "0.1.0"
 edition = "2024"
+publish = false
 
 [[bin]]
 name = "analyze-trajectories"
 path = "src/main.rs"
 
 [dependencies]
+archestra-bench-core = { path = "../core" }
 nitpicker-agent = "0.1"
 rig-core = "0.38"
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
