+++ archestra-bench/core/Cargo.toml
@@ -7,4 +7,3 @@ publish = false
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-toml = "0.8"
