+++ archestra-bench/lanes.toml
@@ -8,15 +8,25 @@
 #   base_url     optional endpoint override (e.g. an Anthropic-compatible gateway)
 #   api_key_env  optional env var holding the key (default: <PROVIDER>_API_KEY)
 
-[[lane]]
-name = "gemini-flash"
-provider = "gemini"
-model = "gemini-3.5-flash"
-
 # [[lane]]
-# name = "or-free"
-# provider = "openrouter"
-# model = "openrouter/free"
+# name = "gemini-flash"
+# provider = "gemini"
+# model = "gemini-3.5-flash"
+
+[[lane]]
+name = "or-free"
+provider = "openrouter"
+model = "openrouter/free"
+
+[[lane]]
+name = "or-poolside"
+provider = "openrouter"
+model = "poolside/laguna-m.1:free"
+
+[[lane]]
+name = "or-owl"
+provider = "openrouter"
+model = "openrouter/owl-alpha"
 
 [[lane]]
 name = "minimax"
